feat: Navigation Compose 도입 및 단일 Activity 구조로 전환
