+++ app/src/main/AndroidManifest.xml
@@ -18,17 +18,7 @@
         android:theme="@style/Theme.Shopping"
         tools:targetApi="31">
         <activity
-            android:name=".ui.cart.CartActivity"
-            android:exported="false"
-            android:label="@string/title_activity_cart"
-            android:theme="@style/Theme.Shopping" />
-        <activity
-            android:name=".ui.productDetail.ProductDetailActivity"
-            android:exported="false"
-            android:label="@string/title_activity_product_detail"
-            android:theme="@style/Theme.Shopping" />
-        <activity
-            android:name=".ui.productList.ProductListActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
