+++ app/src/main/java/woowacourse/shopping/ui/MainActivity.kt
@@ -1,0 +1,21 @@
+package woowacourse.shopping.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import woowacourse.shopping.ShoppingApplication
+import woowacourse.shopping.ui.navigation.AppNavHost
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        val appContainer = (application as ShoppingApplication).appContainer
+
+        setContent {
+            AppNavHost(appContainer = appContainer)
+        }
+    }
+}
+
