--- app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -1,40 +1,0 @@
-package woowacourse.shopping.ui.cart
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.ui.Modifier
-import androidx.lifecycle.viewmodel.compose.viewModel
-import woowacourse.shopping.ShoppingApplication
-
-class CartActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        val appContainer = (application as ShoppingApplication).appContainer
-        setContent {
-            val viewModel: CartViewModel =
-                viewModel(
-                    factory =
-                        CartViewModel.factory(
-                            appContainer.cartRepository,
-                            appContainer.recentProductRepository,
-                            appContainer.productRepository,
-                        ),
-                )
-            Scaffold(
-                modifier = Modifier.fillMaxSize(),
-            ) { innerPadding ->
-                CartScreen(
-                    viewModel = viewModel,
-                    modifier = Modifier.padding(innerPadding),
-                    onClickClose = { finish() },
-                )
-            }
-        }
-    }
-}
